# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "TiendaPOS API",
        "description": "SaaS POS Multi-País para LATAM. Soporta 9 países (VE, CO, MX, EC, AR, PE, CL, BO, UY) con reglas fiscales, IVA, monedas y métodos de pago locales.",
        "contact": {
            "email": "soporte@tiendapos.com"
        },
        "version": "2.1.0"
    },
    "servers": [
        {
            "url": "/api/v1",
            "description": "API Server"
        }
    ],
    "paths": {
...
        "/ordenes-compra": {
            "get": {
                "tags": [
                    "Compras"
                ],
                "summary": "Listar órdenes de compra",
                "operationId": "9ac1172d9995ec4c1220cc02e7bd65ee",
                "parameters": [
                    {
                        "name": "estado",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "proveedor_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Órdenes paginadas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Compras"
                ],
                "summary": "Crear orden de compra",
                "operationId": "f3794db27f1e5a7c31157610f12d3f0e",
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "proveedor_id": {
                                        "type": "integer"
                                    },
                                    "almacen_id": {
                                        "type": "integer"
                                    },
                                    "moneda": {
                                        "type": "string"
                                    },
                                    "items": {
                                        "type": "array",
                                        "items": {
                                            "properties": {
                                                "producto_id": {
                                                    "type": "integer"
                                                },
                                                "cantidad": {
                                                    "type": "number"
                                                },
                                                "costo_unitario": {
                                                    "type": "number"
                                                },
                                                "impuesto_pct": {
                                                    "type": "number"
                                                }
                                            },
                                            "type": "object"
                                        }
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Orden creada"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/ordenes-compra/{id}": {
            "get": {
                "tags": [
                    "Compras"
                ],
                "summary": "Mostrar orden de compra",
                "operationId": "a790af66d21817c9313bc45facac31aa",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orden con items"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Compras"
                ],
                "summary": "Actualizar orden de compra",
                "operationId": "1eca7150634075999420ecad0d2fbf78",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "estado": {
                                        "type": "string",
                                        "enum": [
                                            "borrador",
                                            "aprobada",
                                            "cancelada"
                                        ]
                                    },
                                    "fecha_estimada": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "notas": {
                                        "type": "string"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Orden actualizada"
                    },
                    "422": {
                        "description": "Estado no editable"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Compras"
                ],
                "summary": "Eliminar orden (solo borrador)",
                "operationId": "c52b1634219d8506c563fd1ec5434b15",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Orden eliminada"
                    },
                    "422": {
                        "description": "Solo borradores eliminables"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
        "/sesiones-caja": {
            "get": {
                "tags": [
                    "Cajas"
                ],
                "summary": "Listar sesiones de caja",
                "operationId": "8b0fdabbab4afcc33f5975974ab3a839",
                "parameters": [
                    {
                        "name": "estado",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "caja_id",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    },
                    {
                        "name": "per_page",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sesiones paginadas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/sesiones-caja/{id}": {
            "get": {
                "tags": [
                    "Cajas"
                ],
                "summary": "Mostrar sesión de caja con movimientos y ventas",
                "operationId": "42016fec52028069ea2e3c0b7a64be1f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Sesión con movimientos y ventas"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "Sanctum token de autenticación. Obténlo via POST /api/v1/auth/login",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}